Added a hack to get the cookie for the non-conforming authentication method on lists.w3.org.
When looking for an agenda in a mail archive page, ignore the
navigation links and and only look at the original mail body.

Add the beginnings of support for agendas in HTML (from other sources
than the mail archives).

Avoid printing the results from searching in the background if the bot
already left the channel.

The irc:/ircs: URL passed on the command line may include a channel
name to join.


git-svn-id: file:///home/bert/Tmp/tmpsvnrepo/trunk@33 25f05dee-09ec-4f49-8be2-8cc43a152504