Enhancement for quoted emails and links in calendars in HTML.
Fix: Look for <pre id=body> rather than <pre class=body> in the mail
archives.

Fix: Avoid error message about an undefined variable if no calendars
are associated with a channel.

Enhancement: Add support for agendas in UL or OL lists in HTML, so
that links in such lists can be recognized and output as Ivan
links.:(All agenda parsers now have an extra parameter: the base URI.)

Enhancement: Support lists numbered with "1-", "2-", etc. (E.g.:
https://lists.w3.org/Archives/Public/public-epub-wg/2022Feb/0000.html)

Enhancement: Allow agendas in quoted text in emails (quoted with ">"),
as long as the message was directly pointed at ("agenda:") or does not
have "Re:" in the subject line.