Added an option for testing without an IRC server and some refinements
HTML::FormatText does not deal well with PRE elements that contain
inline elements. Replaced it with my own formatter.

The new formatter also formats A elements as Ivan links ("-> anchor
URL"), so that the URL is visible in IRC and will become a link again
in the formatted minutes.

Added option -o for testing: With -o, agendabot does not connect to an
IRC server, but extracts agendas from URLs given on the command line
and outputs those to a file (or to stdout).

Made some refinements in the heuristic parsers.