Try two_level_agenda_parser() before html_list_agenda_parser().
Only check the list of security exceptions until the first match, and
only do one regexp match per item, for a tiny gain in efficiency.

When extracting the agenda section from a calendar event, include the
<h2>Agenda</h2> at the start of the agenda, because some of the
heuristic parsers look for the word Agenda.

Correctly calculate the earliest possible timestamp ($oldest) when the
request to suggest agenda items includes a period in weeks.

Use the word ‘and’ in the response to a status request only if the
list of calendars to report is not empty.

In two_level_agenda_parser(), also allow lists that are numbered as
‘(1)’, ‘(2)’, etc, not just ‘1)’, ‘2)’.

Also in two_level_agenda_parser(), the indent of a numbered list may
become less if the numbers increase: E.g., ‘10.’ is probably indented
one space less than ‘9.’.

Write messages to the standard error output (which includes log
messages) in UTF-8.